auth/requirements.txt
orjson==3.4.1
uvloop==0.14.0
flasgger==0.9.5
Flask-Migrate==3.1.0
flask==2.2.2
flask-sqlalchemy==2.5.1
gevent==21.12.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
pydantic==1.9.2
SQLAlchemy~=1.4.40
aniso8601==9.0.1
click==8.1.3
Flask-JWT-Extended==4.4.4
Flask-RESTful==0.3.9
greenlet==1.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
passlib==1.7.4
PyJWT==2.4.0
pytz==2022.2.1
six==1.16.0
Werkzeug==2.2.2
zipp==3.8.1
redis==4.3.4
types-redis==4.3.17